Update and rename src/main/java/br/unb/cic/bionimbus/services/storage/StoragePolicy.java to src/main/java/br/unb/cic/bionimbus/services/storage/politcy/StoragePolicy.java
